Check PWD exists when asserting '.' exists
On Unix '.' resolves to a file handle to the directory that was created
whenever the user ran cd. This mean that if the directory is removed '.'
refers to a file handle to a deleted directory, which is not very useful
as you can't do much with it but since the file handle still has the
inode you can stat it so it "exists".

Other relative paths don't suffer this issue, as children won't exist,
and parents are still accessible.

Fixes #24